Refactor laplacian/testcases and API snake_case
Reorganise laplacian interpolation and testcases into standalone packages, and convert public API and internals to snake_case. Moves: laplacian/* → new laplacian_interp package and testcases/* → new testcases package; removed old dvfopt.laplacian and dvfopt.testcases modules. Rename/compat changes: sliceToSlice3DLaplacian → slice_to_slice_3d_laplacian, objectiveEuc → objective_euc, methodName → method_name and other camelCase identifiers updated across core/ modules and notebooks. Additions and behavior changes: introduce _adaptive_maxiter to scale SLSQP iteration budgets with variable count and use it in serial/parallel/3D solvers; guard against non-finite optimizer outputs; make ProcessPoolExecutor.shutdown wait for workers. I/O and utilities: loadNiiImages → load_nii_images and parameter renames; generate_random_dvf now returns float64; simplify dvfopt.io exports and update docs/README/CLAUDE and notebooks to point at the new packages and names. Misc: tidy variable naming and small refactors in jacobian/intersection and other modules to improve clarity.

+++ dvfopt/dvf/generation.py
@@ -23,7 +23,7 @@ def generate_random_dvf(shape, max_magnitude=5.0, seed=None):
     rng = np.random.default_rng(seed)
     C, _, H, W = shape
     assert C == 3, "DVF must have 3 channels (dz, dy, dx)"
-    return rng.uniform(-max_magnitude, max_magnitude, size=shape).astype(np.float32)
+    return rng.uniform(-max_magnitude, max_magnitude, size=shape).astype(np.float64)
 
 
 def generate_random_dvf_3d(shape, max_magnitude=5.0, seed=None):
